Собрал проект после перехода с начального репозитория и обновления с другими разработчиками
