+++ src/containers/Auth/AuthForm.js
@@ -1,9 +1,11 @@
 import React from 'react';
+import { Link } from 'react-router-dom';
 
 const authForm = () => {
     return (
         <div>
-            Форма авторизации
+            <h2>Форма авторизации</h2>
+            <Link to='/register'>Форма регистрации </Link>
         </div>
     );
 }
