+++ src/pages/Register.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-// import RegForm from '../containers/Register/RegForm';
+import RegForm from '../containers/Register/RegForm';
 import Modal from '../components/UI/Modal/Modal';
 import Backdrop from '../components/UI/Backdrop/Backdrop';
 
@@ -10,7 +10,7 @@ const register = () => {
             <Backdrop/>
             <Modal show>
                 <div>Форма регистрации</div>
-                {/* <RegForm/> */}
+                <RegForm/>
             </Modal>
         </>
     );
